feat: Refactor code to separate functions for creating review comments, analyzing code, getting PR details, getting AI response, creating comments, and getting diffs.
