+++ src/analyzeCode.ts
@@ -1,0 +1,29 @@
+import {File} from "parse-diff";
+import {createPrompt, PRDetails} from "./createPrompt";
+import {getAIResponse} from "./getAIResponse";
+import {createComment} from "./createComment";
+
+export async function analyzeCode(
+  parsedDiff: File[],
+  prDetails: PRDetails
+): Promise<Array<{ body: string; path: string; line: number }>> {
+  const comments: Array<{ body: string; path: string; line: number }> = [];
+
+  for (const file of parsedDiff) {
+    if (file.to === "/dev/null") continue; // Ignore deleted files
+    for (const chunk of file.chunks) {
+      // Ignore diffs longer than 2000 characters
+      if (chunk.content.length > 4000) continue;
+
+      const prompt = createPrompt(file, chunk, prDetails);
+      const aiResponse = await getAIResponse(prompt);
+      if (aiResponse) {
+        const newComments = createComment(file, chunk, aiResponse);
+        if (newComments) {
+          comments.push(...newComments);
+        }
+      }
+    }
+  }
+  return comments;
+}
